# swagger_output.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "Subject System REST API",
    "description": ""
  },
...
  "securityDefinitions": {},
  "paths": {
    "/api/studentcourses/": {
      "post": {
        "tags": [
          "StudentCourse"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "studentId": {
                    "example": "any"
                  },
                  "courseId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "StudentCourse"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "studentId": {
                    "example": "any"
                  },
                  "courseId": {
                    "example": "any"
                  },
                  "newCourseId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/studentcourses/{c_id}": {
      "delete": {
        "tags": [
          "StudentCourse"
        ],
        "description": "",
        "parameters": [
          {
            "name": "c_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "studentId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/": {
      "post": {
        "tags": [
          "Course"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "className": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Course"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/courses/{c_id}": {
      "get": {
        "tags": [
          "Course"
        ],
        "description": "",
        "parameters": [
          {
            "name": "c_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Course"
        ],
        "description": "",
        "parameters": [
          {
            "name": "c_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Course"
        ],
        "description": "",
        "parameters": [
          {
            "name": "c_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/courses/{c_id}/students": {
      "get": {
        "tags": [
          "Course"
        ],
        "description": "",
        "parameters": [
          {
            "name": "c_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/teachers/{t_id}/courses": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "description": "",
        "parameters": [
          {
            "name": "t_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/students/": {
      "post": {
        "tags": [
          "Student"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "account": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "address": {
                    "example": "any"
                  },
                  "avatarImage": {
                    "example": "any"
                  },
                  "learning": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Student"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/students/{s_id}": {
      "get": {
        "tags": [
          "Student"
        ],
        "description": "",
        "parameters": [
          {
            "name": "s_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Student"
        ],
        "description": "",
        "parameters": [
          {
            "name": "s_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "account": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Student"
        ],
        "description": "",
        "parameters": [
          {
            "name": "s_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/students/{s_id}/courses": {
      "get": {
        "tags": [
          "Student"
        ],
        "description": "",
        "parameters": [
          {
            "name": "s_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/departments/": {
      "post": {
        "tags": [
          "Department"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Department"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/departments/{d_id}": {
      "get": {
        "tags": [
          "Department"
        ],
        "description": "",
        "parameters": [
          {
            "name": "d_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Department"
        ],
        "description": "",
        "parameters": [
          {
            "name": "d_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Department"
        ],
        "description": "",
        "parameters": [
          {
            "name": "d_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
...
}